Eliminate synchronization conflicts between tsm and tsm2, and between iterator v1 and iterator v2.
    FieldVal / ColumnType refactoring to logical type. Distinguish between physical types of file storage.
