+++ tskv/src/compaction/compact.rs
@@ -13,17 +13,17 @@ use crate::context::GlobalContext;
 use crate::error::Result;
 use crate::summary::{CompactMeta, VersionEdit};
 use crate::tseries_family::TseriesFamily;
-use crate::tsm2::page::{Chunk, ColumnGroup, Page};
-use crate::tsm2::reader::{decode_pages, decode_pages_buf, TSM2MetaData, TSM2Reader};
-use crate::tsm2::writer::{DataBlock2, Tsm2Writer};
-use crate::tsm2::ColumnGroupID;
+use crate::tsm::page::{Chunk, ColumnGroup, Page};
+use crate::tsm::reader::{decode_pages, decode_pages_buf, TsmMetaData, TsmReader};
+use crate::tsm::writer::{DataBlock, TsmWriter};
+use crate::tsm::ColumnGroupID;
 use crate::{ColumnFileId, Error, LevelId, TseriesFamilyId};
 
 /// Temporary compacting data block meta
 #[derive(Clone)]
 pub(crate) struct CompactingBlockMeta {
     reader_idx: usize,
-    reader: Arc<TSM2Reader>,
+    reader: Arc<TsmReader>,
     meta: Arc<Chunk>,
     column_group_id: ColumnGroupID,
 }
@@ -66,7 +66,7 @@ impl Ord for CompactingBlockMeta {
 impl CompactingBlockMeta {
     pub fn new(
         tsm_reader_idx: usize,
-        tsm_reader: Arc<TSM2Reader>,
+        tsm_reader: Arc<TsmReader>,
         chunk: Arc<Chunk>,
         column_group_id: ColumnGroupID,
     ) -> Self {
@@ -92,7 +92,7 @@ impl CompactingBlockMeta {
         Ok(*column_group.time_range())
     }
 
-    pub async fn get_data_block_filter_by_tomb(&self) -> Result<DataBlock2> {
+    pub async fn get_data_block_filter_by_tomb(&self) -> Result<DataBlock> {
         let sid = self.meta.series_id();
         let mut data_block = self
             .reader
@@ -111,7 +111,7 @@ impl CompactingBlockMeta {
             .await
     }
 
-    pub fn tsm_meta(&self) -> Arc<TSM2MetaData> {
+    pub fn tsm_meta(&self) -> Arc<TsmMetaData> {
         self.reader.tsm_meta_data()
     }
 
@@ -258,7 +258,7 @@ impl CompactingBlockMetaGroup {
 
     fn chunk_merged_block(
         &self,
-        data_block: DataBlock2,
+        data_block: DataBlock,
         max_block_size: usize,
     ) -> Result<Vec<CompactingBlock>> {
         let mut merged_blks = Vec::new();
@@ -331,7 +331,7 @@ pub enum CompactingBlock {
         priority: usize,
         series_id: SeriesId,
         series_key: SeriesKey,
-        data_block: DataBlock2,
+        data_block: DataBlock,
     },
     Encoded {
         priority: usize,
@@ -355,7 +355,7 @@ impl CompactingBlock {
         priority: usize,
         series_id: SeriesId,
         series_key: SeriesKey,
-        data_block: DataBlock2,
+        data_block: DataBlock,
     ) -> CompactingBlock {
         Self::Decoded {
             priority,
@@ -399,7 +399,7 @@ impl CompactingBlock {
         }
     }
 
-    pub fn decode(self) -> Result<DataBlock2> {
+    pub fn decode(self) -> Result<DataBlock> {
         match self {
             CompactingBlock::Decoded { data_block, .. } => Ok(data_block),
             CompactingBlock::Encoded {
@@ -448,13 +448,13 @@ impl CompactingBlock {
 
 struct CompactingFile {
     i: usize,
-    tsm_reader: Arc<TSM2Reader>,
+    tsm_reader: Arc<TsmReader>,
     series_idx: usize,
     series_ids: Vec<SeriesId>,
 }
 
 impl CompactingFile {
-    fn new(i: usize, tsm_reader: Arc<TSM2Reader>) -> Self {
+    fn new(i: usize, tsm_reader: Arc<TsmReader>) -> Self {
         let mut series_ids = {
             let chunks = tsm_reader.chunk_group();
             chunks
@@ -514,7 +514,7 @@ impl PartialOrd for CompactingFile {
 }
 
 pub(crate) struct CompactIterator {
-    tsm_readers: Vec<Arc<TSM2Reader>>,
+    tsm_readers: Vec<Arc<TsmReader>>,
     compacting_files: BinaryHeap<Pin<Box<CompactingFile>>>,
     /// Maximum values in generated CompactingBlock
     max_data_block_size: usize,
@@ -558,7 +558,7 @@ impl Default for CompactIterator {
 
 impl CompactIterator {
     pub(crate) fn new(
-        tsm_readers: Vec<Arc<TSM2Reader>>,
+        tsm_readers: Vec<Arc<TsmReader>>,
         max_data_block_size: usize,
         decode_non_overlap_blocks: bool,
     ) -> Self {
@@ -776,10 +776,7 @@ pub async fn run_compaction_job(
     let tsf_id = request.ts_family_id;
     let mut tsm_readers = Vec::new();
     for col_file in request.files.iter() {
-        let tsm_reader = request
-            .version
-            .get_tsm_reader2(col_file.file_path())
-            .await?;
+        let tsm_reader = request.version.get_tsm_reader(col_file.file_path()).await?;
         tsm_readers.push(tsm_reader);
     }
 
@@ -788,7 +785,7 @@ pub async fn run_compaction_job(
     let tsm_dir = request.storage_opt.tsm_dir(&request.database, tsf_id);
     let max_file_size = request.storage_opt.level_max_file_size(request.out_level);
     let mut tsm_writer =
-        Tsm2Writer::open(&tsm_dir, kernel.file_id_next(), max_file_size, false).await?;
+        TsmWriter::open(&tsm_dir, kernel.file_id_next(), max_file_size, false).await?;
     // let mut tsm_writer = tsm::new_tsm_writer(&tsm_dir, kernel.file_id_next(), false, 0).await?;
     info!(
         "Compaction: File: {} been created (level: {}).",
@@ -819,7 +816,7 @@ pub async fn run_compaction_job(
                 .await?
                 {
                     tsm_writer =
-                        Tsm2Writer::open(&tsm_dir, kernel.file_id_next(), max_file_size, false)
+                        TsmWriter::open(&tsm_dir, kernel.file_id_next(), max_file_size, false)
                             .await?;
                 }
             }
@@ -853,7 +850,7 @@ pub async fn run_compaction_job(
             .await?
             {
                 tsm_writer =
-                    Tsm2Writer::open(&tsm_dir, kernel.file_id_next(), max_file_size, false).await?;
+                    TsmWriter::open(&tsm_dir, kernel.file_id_next(), max_file_size, false).await?;
             }
         }
     }
@@ -891,7 +888,7 @@ pub async fn run_compaction_job(
 }
 
 async fn handle_finish_write_tsm_meta(
-    tsm_writer: &mut Tsm2Writer,
+    tsm_writer: &mut TsmWriter,
     file_metas: &mut HashMap<ColumnFileId, Arc<BloomFilter>>,
     version_edit: &mut VersionEdit,
     request: &CompactReq,
@@ -919,7 +916,7 @@ async fn handle_finish_write_tsm_meta(
 }
 
 fn new_compact_meta(
-    tsm_writer: &Tsm2Writer,
+    tsm_writer: &TsmWriter,
     tsf_id: TseriesFamilyId,
     level: LevelId,
 ) -> CompactMeta {
@@ -946,8 +943,8 @@ pub mod test {
     use models::codec::Encoding;
     use models::field_value::FieldVal;
     use models::predicate::domain::TimeRange;
-    use models::schema::{ColumnType, TableColumn, TskvTableSchema};
-    use models::{SeriesId, SeriesKey, ValueType};
+    use models::schema::{ColumnType, PhysicalCType, TableColumn, TskvTableSchema};
+    use models::{PhysicalDType, SeriesId, SeriesKey, ValueType};
 
     use crate::compaction::{run_compaction_job, CompactReq};
     use crate::context::GlobalContext;
@@ -956,13 +953,13 @@ pub mod test {
     use crate::kv_option::Options;
     use crate::summary::VersionEdit;
     use crate::tseries_family::{ColumnFile, LevelInfo, Version};
+    use crate::tsm::reader::TsmReader;
+    use crate::tsm::writer::{Column, DataBlock, TsmWriter};
     use crate::tsm::TsmTombstone;
-    use crate::tsm2::reader::TSM2Reader;
-    use crate::tsm2::writer::{Column, DataBlock2, Tsm2Writer};
 
     pub(crate) async fn write_data_blocks_to_column_file(
         dir: impl AsRef<Path>,
-        data: Vec<HashMap<SeriesId, DataBlock2>>,
+        data: Vec<HashMap<SeriesId, DataBlock>>,
     ) -> (u64, Vec<Arc<ColumnFile>>) {
         if !file_manager::try_exists(&dir) {
             std::fs::create_dir_all(&dir).unwrap();
@@ -971,7 +968,7 @@ pub mod test {
         let mut file_seq = 0;
         for (i, d) in data.iter().enumerate() {
             file_seq = i as u64 + 1;
-            let mut writer = Tsm2Writer::open(&dir, file_seq, 0, false).await.unwrap();
+            let mut writer = TsmWriter::open(&dir, file_seq, 0, false).await.unwrap();
             for (sid, data_blks) in d.iter() {
                 writer
                     .write_datablock(*sid, SeriesKey::default(), data_blks.clone())
@@ -995,8 +992,8 @@ pub mod test {
 
     async fn read_data_blocks_from_column_file(
         path: impl AsRef<Path>,
-    ) -> HashMap<SeriesId, Vec<DataBlock2>> {
-        let tsm_reader = TSM2Reader::open(&path).await.unwrap();
+    ) -> HashMap<SeriesId, Vec<DataBlock>> {
+        let tsm_reader = TsmReader::open(&path).await.unwrap();
         let mut data = HashMap::new();
         for (sid, chunk) in tsm_reader.chunk() {
             let mut blks = vec![];
@@ -1014,7 +1011,8 @@ pub mod test {
 
     fn i64_column(data: Vec<i64>) -> Column {
         let mut col =
-            Column::empty_with_cap(ColumnType::Field(ValueType::Integer), data.len()).unwrap();
+            Column::empty_with_cap(PhysicalCType::Field(PhysicalDType::Integer), data.len())
+                .unwrap();
         for datum in data {
             col.push(Some(FieldVal::Integer(datum)))
         }
@@ -1023,7 +1021,7 @@ pub mod test {
 
     fn ts_column(data: Vec<i64>) -> Column {
         let mut col =
-            Column::empty_with_cap(ColumnType::Time(TimeUnit::Nanosecond), data.len()).unwrap();
+            Column::empty_with_cap(PhysicalCType::Time(TimeUnit::Nanosecond), data.len()).unwrap();
         for datum in data {
             col.push(Some(FieldVal::Integer(datum)))
         }
@@ -1032,7 +1030,8 @@ pub mod test {
 
     fn i64_some_column(data: Vec<Option<i64>>) -> Column {
         let mut col =
-            Column::empty_with_cap(ColumnType::Field(ValueType::Integer), data.len()).unwrap();
+            Column::empty_with_cap(PhysicalCType::Field(PhysicalDType::Integer), data.len())
+                .unwrap();
         for datum in data {
             col.push(datum.map(FieldVal::Integer))
         }
@@ -1041,7 +1040,8 @@ pub mod test {
 
     fn u64_some_column(data: Vec<Option<u64>>) -> Column {
         let mut col =
-            Column::empty_with_cap(ColumnType::Field(ValueType::Unsigned), data.len()).unwrap();
+            Column::empty_with_cap(PhysicalCType::Field(PhysicalDType::Unsigned), data.len())
+                .unwrap();
         for datum in data {
             col.push(datum.map(FieldVal::Unsigned))
         }
@@ -1050,7 +1050,7 @@ pub mod test {
 
     fn f64_some_column(data: Vec<Option<f64>>) -> Column {
         let mut col =
-            Column::empty_with_cap(ColumnType::Field(ValueType::Float), data.len()).unwrap();
+            Column::empty_with_cap(PhysicalCType::Field(PhysicalDType::Float), data.len()).unwrap();
         for datum in data {
             col.push(datum.map(FieldVal::Float))
         }
@@ -1059,7 +1059,8 @@ pub mod test {
 
     fn bool_some_column(data: Vec<Option<bool>>) -> Column {
         let mut col =
-            Column::empty_with_cap(ColumnType::Field(ValueType::Boolean), data.len()).unwrap();
+            Column::empty_with_cap(PhysicalCType::Field(PhysicalDType::Boolean), data.len())
+                .unwrap();
         for datum in data {
             col.push(datum.map(FieldVal::Boolean))
         }
@@ -1079,7 +1080,7 @@ pub mod test {
     async fn check_column_file(
         dir: impl AsRef<Path>,
         version_edit: VersionEdit,
-        expected_data: HashMap<SeriesId, Vec<DataBlock2>>,
+        expected_data: HashMap<SeriesId, Vec<DataBlock>>,
     ) {
         let path = get_result_file_path(dir, version_edit);
         let mut data = read_data_blocks_from_column_file(path).await;
@@ -1175,7 +1176,7 @@ pub mod test {
             ],
         );
         let schema = Arc::new(schema);
-        let data1 = DataBlock2::new(
+        let data1 = DataBlock::new(
             schema.clone(),
             ts_column(vec![1, 2, 3]),
             schema.time_column(),
@@ -1191,7 +1192,7 @@ pub mod test {
             ],
         );
 
-        let data2 = DataBlock2::new(
+        let data2 = DataBlock::new(
             schema.clone(),
             ts_column(vec![4, 5, 6]),
             schema.time_column(),
@@ -1207,7 +1208,7 @@ pub mod test {
             ],
         );
 
-        let data3 = DataBlock2::new(
+        let data3 = DataBlock::new(
             schema.clone(),
             ts_column(vec![7, 8, 9]),
             schema.time_column(),
@@ -1223,7 +1224,7 @@ pub mod test {
             ],
         );
 
-        let expected_data = DataBlock2::new(
+        let expected_data = DataBlock::new(
             schema.clone(),
             ts_column(vec![1, 2, 3, 4, 5, 6, 7, 8, 9]),
             schema.time_column(),
@@ -1296,7 +1297,7 @@ pub mod test {
             ],
         );
         let schema = Arc::new(schema);
-        let data1 = DataBlock2::new(
+        let data1 = DataBlock::new(
             schema.clone(),
             ts_column(vec![4, 5, 6]),
             schema.time_column(),
@@ -1312,7 +1313,7 @@ pub mod test {
             ],
         );
 
-        let data2 = DataBlock2::new(
+        let data2 = DataBlock::new(
             schema.clone(),
             ts_column(vec![1, 2, 3]),
             schema.time_column(),
@@ -1328,7 +1329,7 @@ pub mod test {
             ],
         );
 
-        let data3 = DataBlock2::new(
+        let data3 = DataBlock::new(
             schema.clone(),
             ts_column(vec![7, 8, 9]),
             schema.time_column(),
@@ -1344,7 +1345,7 @@ pub mod test {
             ],
         );
 
-        let expected_data = DataBlock2::new(
+        let expected_data = DataBlock::new(
             schema.clone(),
             ts_column(vec![1, 2, 3, 4, 5, 6, 7, 8, 9]),
             schema.time_column(),
@@ -1423,7 +1424,7 @@ pub mod test {
             ],
         );
         let schema = Arc::new(schema);
-        let data1 = DataBlock2::new(
+        let data1 = DataBlock::new(
             schema.clone(),
             ts_column(vec![1, 2, 3, 4]),
             schema.time_column(),
@@ -1439,7 +1440,7 @@ pub mod test {
             ],
         );
 
-        let data2 = DataBlock2::new(
+        let data2 = DataBlock::new(
             schema.clone(),
             ts_column(vec![4, 5, 6, 7]),
             schema.time_column(),
@@ -1455,7 +1456,7 @@ pub mod test {
             ],
         );
 
-        let data3 = DataBlock2::new(
+        let data3 = DataBlock::new(
             schema.clone(),
             ts_column(vec![7, 8, 9]),
             schema.time_column(),
@@ -1471,7 +1472,7 @@ pub mod test {
             ],
         );
 
-        let expected_data = DataBlock2::new(
+        let expected_data = DataBlock::new(
             schema.clone(),
             ts_column(vec![1, 2, 3, 4, 5, 6, 7, 8, 9]),
             schema.time_column(),
@@ -1534,7 +1535,7 @@ pub mod test {
 
     fn generate_column_ts(min_ts: i64, max_ts: i64) -> Column {
         let mut col = Column::empty_with_cap(
-            ColumnType::Time(TimeUnit::Nanosecond),
+            PhysicalCType::Time(TimeUnit::Nanosecond),
             (max_ts - min_ts + 1) as usize,
         )
         .unwrap();
@@ -1545,7 +1546,8 @@ pub mod test {
     }
 
     fn generate_column_i64(len: usize, none_range: Vec<(usize, usize)>) -> Column {
-        let mut col = Column::empty_with_cap(ColumnType::Field(ValueType::Integer), len).unwrap();
+        let mut col =
+            Column::empty_with_cap(PhysicalCType::Field(PhysicalDType::Integer), len).unwrap();
         for i in 0..len {
             if none_range.iter().any(|(min, max)| i >= *min && i <= *max) {
                 col.push(None);
@@ -1557,7 +1559,8 @@ pub mod test {
     }
 
     fn generate_column_u64(len: usize, none_range: Vec<(usize, usize)>) -> Column {
-        let mut col = Column::empty_with_cap(ColumnType::Field(ValueType::Unsigned), len).unwrap();
+        let mut col =
+            Column::empty_with_cap(PhysicalCType::Field(PhysicalDType::Unsigned), len).unwrap();
         for i in 0..len {
             if none_range.iter().any(|(min, max)| i >= *min && i <= *max) {
                 col.push(None);
@@ -1569,7 +1572,8 @@ pub mod test {
     }
 
     fn generate_column_f64(len: usize, none_range: Vec<(usize, usize)>) -> Column {
-        let mut col = Column::empty_with_cap(ColumnType::Field(ValueType::Float), len).unwrap();
+        let mut col =
+            Column::empty_with_cap(PhysicalCType::Field(PhysicalDType::Float), len).unwrap();
         for i in 0..len {
             if none_range.iter().any(|(min, max)| i >= *min && i <= *max) {
                 col.push(None);
@@ -1581,7 +1585,8 @@ pub mod test {
     }
 
     fn generate_column_bool(len: usize, none_range: Vec<(usize, usize)>) -> Column {
-        let mut col = Column::empty_with_cap(ColumnType::Field(ValueType::Boolean), len).unwrap();
+        let mut col =
+            Column::empty_with_cap(PhysicalCType::Field(PhysicalDType::Boolean), len).unwrap();
         for i in 0..len {
             if none_range.iter().any(|(min, max)| i >= *min && i <= *max) {
                 col.push(None);
@@ -1641,7 +1646,7 @@ pub mod test {
             (
                 1_u64,
                 vec![
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema1.clone(),
                         generate_column_ts(1, 1000),
                         schema1.time_column(),
@@ -1652,7 +1657,7 @@ pub mod test {
                         ],
                         schema1.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema1.clone(),
                         generate_column_ts(1001, 2000),
                         schema1.time_column(),
@@ -1663,7 +1668,7 @@ pub mod test {
                         ],
                         schema1.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema1.clone(),
                         generate_column_ts(2001, 2500),
                         schema1.time_column(),
@@ -1679,7 +1684,7 @@ pub mod test {
             (
                 2,
                 vec![
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(1, 1000),
                         schema2.time_column(),
@@ -1691,7 +1696,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(1001, 2000),
                         schema2.time_column(),
@@ -1703,7 +1708,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(2001, 3000),
                         schema2.time_column(),
@@ -1715,7 +1720,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(3001, 4000),
                         schema2.time_column(),
@@ -1727,7 +1732,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(4001, 4500),
                         schema2.time_column(),
@@ -1744,7 +1749,7 @@ pub mod test {
             (
                 3,
                 vec![
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(1001, 2000),
                         schema2.time_column(),
@@ -1756,7 +1761,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(2001, 3000),
                         schema2.time_column(),
@@ -1768,7 +1773,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(3001, 4000),
                         schema2.time_column(),
@@ -1780,7 +1785,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(4001, 5000),
                         schema2.time_column(),
@@ -1792,7 +1797,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(5001, 6000),
                         schema2.time_column(),
@@ -1804,7 +1809,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(6001, 6500),
                         schema2.time_column(),
@@ -1819,8 +1824,8 @@ pub mod test {
                 ],
             ),
         ];
-        let expected_data: Vec<DataBlock2> = vec![
-            DataBlock2::new(
+        let expected_data: Vec<DataBlock> = vec![
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(1, 1000),
                 schema2.time_column(),
@@ -1832,7 +1837,7 @@ pub mod test {
                 ],
                 schema2.fields(),
             ),
-            DataBlock2::new(
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(1001, 2000),
                 schema2.time_column(),
@@ -1844,7 +1849,7 @@ pub mod test {
                 ],
                 schema2.fields(),
             ),
-            DataBlock2::new(
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(2001, 3000),
                 schema2.time_column(),
@@ -1856,7 +1861,7 @@ pub mod test {
                 ],
                 schema2.fields(),
             ),
-            DataBlock2::new(
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(3001, 4000),
                 schema2.time_column(),
@@ -1868,7 +1873,7 @@ pub mod test {
                 ],
                 schema2.fields(),
             ),
-            DataBlock2::new(
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(4001, 5000),
                 schema2.time_column(),
@@ -1880,7 +1885,7 @@ pub mod test {
                 ],
                 schema2.fields(),
             ),
-            DataBlock2::new(
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(5001, 6000),
                 schema2.time_column(),
@@ -1892,7 +1897,7 @@ pub mod test {
                 ],
                 schema2.fields(),
             ),
-            DataBlock2::new(
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(6001, 6500),
                 schema2.time_column(),
@@ -1917,9 +1922,7 @@ pub mod test {
 
         let mut column_files = Vec::new();
         for (tsm_sequence, args) in data_desc.into_iter() {
-            let mut tsm_writer = Tsm2Writer::open(&dir, tsm_sequence, 0, false)
-                .await
-                .unwrap();
+            let mut tsm_writer = TsmWriter::open(&dir, tsm_sequence, 0, false).await.unwrap();
             for arg in args.into_iter() {
                 tsm_writer
                     .write_datablock(1, SeriesKey::default(), arg)
@@ -1999,7 +2002,7 @@ pub mod test {
             (
                 1_u64,
                 vec![
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema1.clone(),
                         generate_column_ts(1, 1000),
                         schema1.time_column(),
@@ -2010,7 +2013,7 @@ pub mod test {
                         ],
                         schema1.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema1.clone(),
                         generate_column_ts(1001, 2000),
                         schema1.time_column(),
@@ -2021,7 +2024,7 @@ pub mod test {
                         ],
                         schema1.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema1.clone(),
                         generate_column_ts(2001, 2500),
                         schema1.time_column(),
@@ -2037,7 +2040,7 @@ pub mod test {
             (
                 2,
                 vec![
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(1, 1000),
                         schema2.time_column(),
@@ -2049,7 +2052,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(1001, 2000),
                         schema2.time_column(),
@@ -2061,7 +2064,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(2001, 3000),
                         schema2.time_column(),
@@ -2073,7 +2076,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(3001, 4000),
                         schema2.time_column(),
@@ -2085,7 +2088,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(4001, 4500),
                         schema2.time_column(),
@@ -2102,7 +2105,7 @@ pub mod test {
             (
                 3,
                 vec![
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(1001, 2000),
                         schema2.time_column(),
@@ -2114,7 +2117,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(2001, 3000),
                         schema2.time_column(),
@@ -2126,7 +2129,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(3001, 4000),
                         schema2.time_column(),
@@ -2138,7 +2141,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(4001, 5000),
                         schema2.time_column(),
@@ -2150,7 +2153,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(5001, 6000),
                         schema2.time_column(),
@@ -2162,7 +2165,7 @@ pub mod test {
                         ],
                         schema2.fields(),
                     ),
-                    DataBlock2::new(
+                    DataBlock::new(
                         schema2.clone(),
                         generate_column_ts(6001, 6500),
                         schema2.time_column(),
@@ -2177,8 +2180,8 @@ pub mod test {
                 ],
             ),
         ];
-        let expected_data: Vec<DataBlock2> = vec![
-            DataBlock2::new(
+        let expected_data: Vec<DataBlock> = vec![
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(1, 1000),
                 schema2.time_column(),
@@ -2190,7 +2193,7 @@ pub mod test {
                 ],
                 schema2.fields(),
             ),
-            DataBlock2::new(
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(1001, 2000),
                 schema2.time_column(),
@@ -2202,7 +2205,7 @@ pub mod test {
                 ],
                 schema2.fields(),
             ),
-            DataBlock2::new(
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(2001, 3000),
                 schema2.time_column(),
@@ -2214,7 +2217,7 @@ pub mod test {
                 ],
                 schema2.fields(),
             ),
-            DataBlock2::new(
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(3001, 4000),
                 schema2.time_column(),
@@ -2226,7 +2229,7 @@ pub mod test {
                 ],
                 schema2.fields(),
             ),
-            DataBlock2::new(
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(4001, 5000),
                 schema2.time_column(),
@@ -2238,7 +2241,7 @@ pub mod test {
                 ],
                 schema2.fields(),
             ),
-            DataBlock2::new(
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(5001, 6000),
                 schema2.time_column(),
@@ -2250,7 +2253,7 @@ pub mod test {
                 ],
                 schema2.fields(),
             ),
-            DataBlock2::new(
+            DataBlock::new(
                 schema2.clone(),
                 generate_column_ts(6001, 6500),
                 schema2.time_column(),
@@ -2275,9 +2278,7 @@ pub mod test {
 
         let mut column_files = Vec::new();
         for (tsm_sequence, args) in data_desc.into_iter() {
-            let mut tsm_writer = Tsm2Writer::open(&dir, tsm_sequence, 0, false)
-                .await
-                .unwrap();
+            let mut tsm_writer = TsmWriter::open(&dir, tsm_sequence, 0, false).await.unwrap();
             for arg in args.into_iter() {
                 tsm_writer
                     .write_datablock(1, SeriesKey::default(), arg)
