+++ tskv/src/memcache.rs
@@ -12,7 +12,7 @@ use minivec::MiniVec;
 use models::field_value::FieldVal;
 use models::predicate::domain::{TimeRange, TimeRanges};
 use models::schema::{
-    timestamp_convert, ColumnType, Precision, TableColumn, TskvTableSchema, TskvTableSchemaRef,
+    timestamp_convert, PhysicalCType, Precision, TableColumn, TskvTableSchema, TskvTableSchemaRef,
 };
 use models::{ColumnId, RwLockRef, SeriesId, SeriesKey, Timestamp};
 use parking_lot::RwLock;
@@ -24,8 +24,8 @@ use utils::bitset::ImmutBitSet;
 use crate::database::FbSchema;
 use crate::error::Result;
 use crate::tseries_family::Version;
-use crate::tsm2::writer::{Column as ColumnData, DataBlock2};
-use crate::tsm2::TsmWriteData;
+use crate::tsm::writer::{Column as ColumnData, DataBlock};
+use crate::tsm::TsmWriteData;
 use crate::{Error, TseriesFamilyId};
 
 // use skiplist::ordered_skiplist::OrderedSkipList;
@@ -513,17 +513,17 @@ impl SeriesData {
     pub fn build_data_block(
         &self,
         version: Arc<Version>,
-    ) -> Result<Option<(String, DataBlock2, DataBlock2)>> {
+    ) -> Result<Option<(String, DataBlock, DataBlock)>> {
         if let Some(schema) = self.get_schema() {
             let field_ids = schema.fields_id();
 
             let mut cols = schema
                 .fields()
                 .iter()
-                .map(|col| ColumnData::empty(col.column_type.clone()))
+                .map(|col| ColumnData::empty(col.column_type.to_physical_type()))
                 .collect::<Result<Vec<_>>>()?;
             let mut delta_cols = cols.clone();
-            let mut time_array = ColumnData::empty(ColumnType::Time(TimeUnit::from(
+            let mut time_array = ColumnData::empty(PhysicalCType::Time(TimeUnit::from(
                 schema.time_column_precision(),
             )))?;
 
@@ -573,14 +573,14 @@ impl SeriesData {
             }
             return Ok(Some((
                 schema.name.clone(),
-                DataBlock2::new(
+                DataBlock::new(
                     schema.clone(),
                     time_array,
                     schema.time_column(),
                     cols,
                     cols_desc.clone(),
                 ),
-                DataBlock2::new(
+                DataBlock::new(
                     schema.clone(),
                     delta_time_array,
                     schema.time_column(),
