+++ tskv/src/reader/table_scan.rs
@@ -9,7 +9,7 @@ use models::meta_data::VnodeId;
 use tokio::runtime::Runtime;
 use trace::SpanRecorder;
 
-use super::{iterator_v2, SendableTskvRecordBatchStream};
+use super::{iterator, SendableTskvRecordBatchStream};
 use crate::reader::QueryOption;
 use crate::{EngineRef, Error};
 
@@ -29,7 +29,7 @@ impl LocalTskvTableScanStream {
         runtime: Arc<Runtime>,
         span_recorder: SpanRecorder,
     ) -> Self {
-        let iter_future = Box::pin(iterator_v2::execute(
+        let iter_future = Box::pin(iterator::execute(
             runtime,
             kv_inst,
             option,
