+++ tskv/src/tsm/writer.rs
@@ -8,8 +8,8 @@ use minivec::MiniVec;
 use models::codec::Encoding;
 use models::field_value::FieldVal;
 use models::predicate::domain::TimeRange;
-use models::schema::{ColumnType, TableColumn, TskvTableSchemaRef};
-use models::{ColumnId, SeriesId, SeriesKey, ValueType};
+use models::schema::{PhysicalCType, TableColumn, TskvTableSchemaRef};
+use models::{ColumnId, PhysicalDType, SeriesId, SeriesKey};
 use snafu::ResultExt;
 use utils::bitset::BitSet;
 use utils::BloomFilter;
@@ -23,12 +23,11 @@ use crate::file_utils::{make_delta_file, make_tsm_file};
 use crate::tsm::codec::{
     get_bool_codec, get_f64_codec, get_i64_codec, get_str_codec, get_u64_codec,
 };
-use crate::tsm::TsmTombstone;
-use crate::tsm2::page::{
+use crate::tsm::page::{
     Chunk, ChunkGroup, ChunkGroupMeta, ChunkGroupWriteSpec, ChunkStatics, ChunkWriteSpec,
     ColumnGroup, Footer, Page, PageMeta, PageStatistics, PageWriteSpec, SeriesMeta, TableMeta,
 };
-use crate::tsm2::{TsmWriteData, BLOOM_FILTER_BITS};
+use crate::tsm::{TsmTombstone, TsmWriteData, BLOOM_FILTER_BITS};
 use crate::{Error, Result};
 
 // #[derive(Debug, Clone)]
@@ -97,29 +96,29 @@ use crate::{Error, Result};
 /// max size 1024
 #[derive(Debug, Clone, PartialEq)]
 pub struct Column {
-    column_type: ColumnType,
+    column_type: PhysicalCType,
     valid: BitSet,
     data: ColumnData,
 }
 
 impl Column {
-    pub fn empty(column_type: ColumnType) -> Result<Column> {
+    pub fn empty(column_type: PhysicalCType) -> Result<Column> {
         let valid = BitSet::new();
         let column = Self {
             column_type: column_type.clone(),
             valid,
-            data: match column_type {
-                ColumnType::Tag => ColumnData::String(vec![], String::new(), String::new()),
-                ColumnType::Time(_) => ColumnData::I64(vec![], i64::MAX, i64::MIN),
-                ColumnType::Field(ref field_type) => match field_type {
-                    ValueType::Float => ColumnData::F64(vec![], f64::MAX, f64::MIN),
-                    ValueType::Integer => ColumnData::I64(vec![], i64::MAX, i64::MIN),
-                    ValueType::Unsigned => ColumnData::U64(vec![], u64::MAX, u64::MIN),
-                    ValueType::Boolean => ColumnData::Bool(vec![], false, true),
-                    ValueType::Geometry(_) | ValueType::String => {
+            data: match column_type.clone() {
+                PhysicalCType::Tag => ColumnData::String(vec![], String::new(), String::new()),
+                PhysicalCType::Time(_) => ColumnData::I64(vec![], i64::MAX, i64::MIN),
+                PhysicalCType::Field(ref field_type) => match field_type {
+                    PhysicalDType::Float => ColumnData::F64(vec![], f64::MAX, f64::MIN),
+                    PhysicalDType::Integer => ColumnData::I64(vec![], i64::MAX, i64::MIN),
+                    PhysicalDType::Unsigned => ColumnData::U64(vec![], u64::MAX, u64::MIN),
+                    PhysicalDType::Boolean => ColumnData::Bool(vec![], false, true),
+                    PhysicalDType::String => {
                         ColumnData::String(vec![], String::new(), String::new())
                     }
-                    ValueType::Unknown => {
+                    PhysicalDType::Unknown => {
                         return Err(Error::UnsupportedDataType {
                             dt: "unknown".to_string(),
                         })
@@ -130,31 +129,35 @@ impl Column {
         Ok(column)
     }
 
-    pub fn empty_with_cap(column_type: ColumnType, cap: usize) -> Result<Column> {
+    pub fn empty_with_cap(column_type: PhysicalCType, cap: usize) -> Result<Column> {
         let valid = BitSet::with_size(cap);
         let column = Self {
             column_type: column_type.clone(),
             valid,
             data: match column_type {
-                ColumnType::Tag => {
+                PhysicalCType::Tag => {
                     ColumnData::String(Vec::with_capacity(cap), String::new(), String::new())
                 }
-                ColumnType::Time(_) => ColumnData::I64(Vec::with_capacity(cap), i64::MAX, i64::MIN),
-                ColumnType::Field(ref field_type) => match field_type {
-                    ValueType::Float => {
+                PhysicalCType::Time(_) => {
+                    ColumnData::I64(Vec::with_capacity(cap), i64::MAX, i64::MIN)
+                }
+                PhysicalCType::Field(ref field_type) => match field_type {
+                    PhysicalDType::Float => {
                         ColumnData::F64(Vec::with_capacity(cap), f64::MAX, f64::MIN)
                     }
-                    ValueType::Integer => {
+                    PhysicalDType::Integer => {
                         ColumnData::I64(Vec::with_capacity(cap), i64::MAX, i64::MIN)
                     }
-                    ValueType::Unsigned => {
+                    PhysicalDType::Unsigned => {
                         ColumnData::U64(Vec::with_capacity(cap), u64::MAX, u64::MIN)
                     }
-                    ValueType::Boolean => ColumnData::Bool(Vec::with_capacity(cap), false, true),
-                    ValueType::Geometry(_) | ValueType::String => {
+                    PhysicalDType::Boolean => {
+                        ColumnData::Bool(Vec::with_capacity(cap), false, true)
+                    }
+                    PhysicalDType::String => {
                         ColumnData::String(Vec::with_capacity(cap), String::new(), String::new())
                     }
-                    ValueType::Unknown => {
+                    PhysicalDType::Unknown => {
                         return Err(Error::UnsupportedDataType {
                             dt: "unknown".to_string(),
                         })
@@ -489,7 +492,7 @@ impl ColumnData {
 }
 
 #[derive(Debug, Clone, PartialEq)]
-pub struct DataBlock2 {
+pub struct DataBlock {
     schema: TskvTableSchemaRef,
     ts: Column,
     ts_desc: TableColumn,
@@ -503,7 +506,7 @@ enum Merge {
     Equal(usize, usize),
 }
 
-impl DataBlock2 {
+impl DataBlock {
     const BLOCK_SIZE: usize = 1000;
 
     pub fn new(
@@ -513,7 +516,7 @@ impl DataBlock2 {
         cols: Vec<Column>,
         cols_desc: Vec<TableColumn>,
     ) -> Self {
-        DataBlock2 {
+        DataBlock {
             schema,
             ts,
             ts_desc,
@@ -535,7 +538,7 @@ impl DataBlock2 {
         Ok(pages)
     }
 
-    pub fn merge(&mut self, other: DataBlock2) -> Result<DataBlock2> {
+    pub fn merge(&mut self, other: DataBlock) -> Result<DataBlock> {
         self.schema_check(&other)?;
 
         let schema = if self.schema.schema_version > other.schema.schema_version {
@@ -548,7 +551,7 @@ impl DataBlock2 {
         let mut columns_des = Vec::new();
         for field in schema.fields() {
             let mut merge_column =
-                Column::empty_with_cap(field.column_type.clone(), time_array.len())?;
+                Column::empty_with_cap(field.column_type.to_physical_type(), time_array.len())?;
             let column_self = self.column(field.id);
             let column_other = other.column(field.id);
             for idx in sort_index.iter() {
@@ -590,13 +593,15 @@ impl DataBlock2 {
             columns_des.push(field.clone());
         }
 
-        let mut ts_col =
-            Column::empty_with_cap(self.ts_desc.column_type.clone(), time_array.len())?;
+        let mut ts_col = Column::empty_with_cap(
+            self.ts_desc.column_type.to_physical_type(),
+            time_array.len(),
+        )?;
 
         time_array
             .iter()
             .for_each(|it| ts_col.push(Some(FieldVal::Integer(*it))));
-        let datablock = DataBlock2::new(schema, ts_col, self.ts_desc.clone(), columns, columns_des);
+        let datablock = DataBlock::new(schema, ts_col, self.ts_desc.clone(), columns, columns_des);
 
         // todo: split datablock to blocks
         // let mut blocks = Vec::with_capacity(time_array.len() / Self::BLOCK_SIZE + 1);
@@ -605,7 +610,7 @@ impl DataBlock2 {
         Ok(datablock)
     }
 
-    fn sort_index_and_time_col(&self, other: &DataBlock2) -> Result<(Vec<Merge>, Vec<i64>)> {
+    fn sort_index_and_time_col(&self, other: &DataBlock) -> Result<(Vec<Merge>, Vec<i64>)> {
         let mut sort_index = Vec::with_capacity(self.len() + other.len());
         let mut time_array = Vec::new();
         let (mut index_self, mut index_other) = (0_usize, 0_usize);
@@ -672,7 +677,7 @@ impl DataBlock2 {
         self.ts.data.is_empty()
     }
 
-    pub fn schema_check(&self, other: &DataBlock2) -> Result<()> {
+    pub fn schema_check(&self, other: &DataBlock) -> Result<()> {
         if self.schema.name != other.schema.name
             || self.schema.db != other.schema.db
             || self.schema.tenant != other.schema.tenant
@@ -701,7 +706,7 @@ impl DataBlock2 {
         None
     }
 
-    pub fn chunk(&self, start: usize, end: usize) -> Result<DataBlock2> {
+    pub fn chunk(&self, start: usize, end: usize) -> Result<DataBlock> {
         if start > end || end > self.len() {
             return Err(Error::CommonError {
                 reason: "start or end index out of range".to_string(),
@@ -714,7 +719,7 @@ impl DataBlock2 {
             .iter()
             .map(|column| column.chunk(start, end))
             .collect::<Result<Vec<_>>>()?;
-        let datablock = DataBlock2::new(
+        let datablock = DataBlock::new(
             self.schema.clone(),
             ts_column,
             self.ts_desc.clone(),
@@ -802,7 +807,7 @@ const HEADER_LEN: u64 = 5;
 const TSM_MAGIC: [u8; 4] = 0x12CDA16_u32.to_be_bytes();
 const VERSION: [u8; 1] = [1];
 
-pub struct Tsm2Writer {
+pub struct TsmWriter {
     file_id: u64,
     min_ts: i64,
     max_ts: i64,
@@ -828,7 +833,7 @@ pub struct Tsm2Writer {
 }
 
 //MutableRecordBatch
-impl Tsm2Writer {
+impl TsmWriter {
     pub async fn open(
         path_buf: &impl AsRef<Path>,
         file_id: u64,
@@ -1016,11 +1021,11 @@ impl Tsm2Writer {
         &mut self,
         series_id: SeriesId,
         series_key: SeriesKey,
-        datablock: DataBlock2,
+        datablock: DataBlock,
     ) -> Result<()> {
         if self.state == State::Finished {
             return Err(Error::CommonError {
-                reason: "Tsm2Writer has been finished".to_string(),
+                reason: "TsmWriter has been finished".to_string(),
             });
         }
 
@@ -1181,14 +1186,14 @@ mod test {
     use models::codec::Encoding;
     use models::field_value::FieldVal;
     use models::predicate::domain::TimeRange;
-    use models::schema::{ColumnType, TableColumn, TskvTableSchema};
-    use models::{SeriesKey, ValueType};
+    use models::schema::{ColumnType, PhysicalCType, TableColumn, TskvTableSchema};
+    use models::{PhysicalDType, SeriesKey, ValueType};
 
-    use crate::tsm2::reader::TSM2Reader;
-    use crate::tsm2::writer::{Column, DataBlock2, Tsm2Writer};
+    use crate::tsm::reader::TsmReader;
+    use crate::tsm::writer::{Column, DataBlock, TsmWriter};
 
     fn i64_column(data: Vec<i64>) -> Column {
-        let mut col = Column::empty(ColumnType::Field(ValueType::Integer)).unwrap();
+        let mut col = Column::empty(PhysicalCType::Field(PhysicalDType::Integer)).unwrap();
         for datum in data {
             col.push(Some(FieldVal::Integer(datum)))
         }
@@ -1196,7 +1201,7 @@ mod test {
     }
 
     fn ts_column(data: Vec<i64>) -> Column {
-        let mut col = Column::empty(ColumnType::Time(TimeUnit::Nanosecond)).unwrap();
+        let mut col = Column::empty(PhysicalCType::Time(TimeUnit::Nanosecond)).unwrap();
         for datum in data {
             col.push(Some(FieldVal::Integer(datum)))
         }
@@ -1237,7 +1242,7 @@ mod test {
             ],
         );
         let schema = Arc::new(schema);
-        let data1 = DataBlock2::new(
+        let data1 = DataBlock::new(
             schema.clone(),
             ts_column(vec![1, 2, 3]),
             schema.time_column(),
@@ -1253,8 +1258,8 @@ mod test {
             ],
         );
 
-        let path = "/tmp/test/tsm2";
-        let mut tsm_writer = Tsm2Writer::open(&PathBuf::from(path), 1, 0, false)
+        let path = "/tmp/test/tsm";
+        let mut tsm_writer = TsmWriter::open(&PathBuf::from(path), 1, 0, false)
             .await
             .unwrap();
         tsm_writer
@@ -1262,7 +1267,7 @@ mod test {
             .await
             .unwrap();
         tsm_writer.finish().await.unwrap();
-        let tsm_reader = TSM2Reader::open(tsm_writer.path).await.unwrap();
+        let tsm_reader = TsmReader::open(tsm_writer.path).await.unwrap();
         let data2 = tsm_reader.read_datablock(1, 0).await.unwrap();
         assert_eq!(data1, data2);
         let time_range = data2.time_range().unwrap();
